backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-SocketIO
python-dotenv
Flask-CORS
requests